Refactored GUI elements
- Created ``panels`` and ``widgets`` modules
- Fixed damage distribution pie chart color fill
- Moved logs widget into a sidebar
