+++ Cargo.toml
@@ -12,6 +12,7 @@ eframe = { version = "0.31.1", features = [
     "default"
 ] }
 egui_plot = "0.31.0"
+egui_material_icons = "0.3.0"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 chrono = { version = "0.4", features = ["serde"] }
@@ -19,4 +20,3 @@ tokio = { version = "1.44.1", features = ["full"] }
 csv = "1.3.1"
 anyhow = "1.0.97"
 directories = "5.0"
-plotters = "0.3.3"
