+++ src/panels.rs
@@ -1,0 +1,5 @@
+pub mod av_panel;
+pub mod menubar_panel;
+pub mod central_panel;
+pub mod statusbar_panel;
+pub mod sidebar_panel;
